changed how the SPA renders html
html pages are stored in variables and are rendered on button onclicks. also made a function that gets all the elements, of the rendered html, that have an id so we can use them later in the code instead of recalling them everytime we need them.
also also,
all the information of the game that is currently being played is stored in an object so we can copy that object and resign it later to its original state when the player finishes the game.